DOC: Shift interactive_view to visualisation.py
